+++ src/test/kotlin/io/github/leanish/gradleconventions/ConfigurationCacheCompatibilityTest.kt
@@ -2,6 +2,7 @@ package io.github.leanish.gradleconventions
 
 import io.github.leanish.gradleconventions.ConventionProperties.BASE_PACKAGE_ENV
 import io.github.leanish.gradleconventions.ConventionProperties.GITHUB_REPOSITORY_OWNER_ENV
+import io.github.leanish.gradleconventions.ConventionProperties.MAVEN_LOCAL_ENABLED_ENV
 import io.github.leanish.gradleconventions.ConventionProperties.MAVEN_CENTRAL_ENABLED_ENV
 import io.github.leanish.gradleconventions.ConventionProperties.PUBLISHING_ENABLED_ENV
 import io.github.leanish.gradleconventions.ConventionProperties.PUBLISHING_GITHUB_OWNER_ENV
@@ -170,6 +171,7 @@ class ConfigurationCacheCompatibilityTest {
 
     private fun testEnvironment(environmentOverrides: Map<String, String>): Map<String, String> {
         return System.getenv().toMutableMap().apply {
+            remove(MAVEN_LOCAL_ENABLED_ENV)
             remove(MAVEN_CENTRAL_ENABLED_ENV)
             remove(PUBLISHING_ENABLED_ENV)
             remove(PUBLISHING_GITHUB_OWNER_ENV)